requirements.txt
# --- Authentic CycleGAN Baseline (2020 era) ---
torch==1.7.1
torchvision==0.8.2
numpy==1.19.5
# Legacy Imaging
Pillow==8.0.0
# Core Visualization & Math
scipy==1.5.4
dominate==2.4.0
visdom==0.1.8.8
tqdm==4.46.1